bix normalized bug in VoxelsLoad.
