+++ source/MRMesh/MRCylinderApproximator.h
@@ -8,21 +8,6 @@
 #include "MRConstants.h"
 
 
-
-#pragma warning(push)
-#pragma warning(disable: 4068) // unknown pragmas
-#pragma warning(disable: 4127) // conditional expression is constant
-#pragma warning(disable: 4464) // relative include path contains '..'
-#pragma warning(disable: 4643) // Forward declaring 'tuple' in namespace std is not permitted by the C++ Standard.
-#pragma warning(disable: 5054) // operator '|': deprecated between enumerations of different types
-#pragma warning(disable: 4244) // casting float to double 
-
-#include <unsupported/Eigen/NonLinearOptimization>
-#include <unsupported/Eigen/NumericalDiff>
-
-#pragma warning(pop)
-
-
 // https://www.geometrictools.com/Documentation/LeastSquaresFitting.pdf 
 
 namespace MR
@@ -384,521 +369,6 @@ private:
         return G( W, PC, resultedRootSquare );
     };
 };
-
-
-
-
-#if 0 
-
-struct PointCloud
-{
-    Eigen::MatrixXd points; // Матрица размерности n x 3, где n - количество точек
-};
-
-
-class ConicSurfaceModel
-{
-public:
-    typedef double Scalar;
-    typedef Eigen::Matrix<Scalar, 3, 1> InputType;
-    typedef Eigen::Matrix<Scalar, 1, 1> ValueType;
-    typedef Eigen::Matrix<Scalar, 1, 3> JacobianType;
-
-
-
-    int operator()( const Eigen::VectorXd& x, Eigen::VectorXd& fvec ) const
-    {
-        const Scalar a = x[0];
-        const Scalar b = x[1];
-        const Scalar c = x[2];
-
-        for ( int i = 0; i < cloud_.points.rows(); ++i )
-        {
-            const InputType p = cloud_.points.row( i ).transpose();
-            const ValueType residual = ( p.transpose() * A_ * p ) / ( p.transpose() * B_ * p ) - Scalar( 1.0 );
-            fvec[i] = residual( 0 );
-        }
-
-        return 0;
-    }
-
-    int df( const Eigen::VectorXd& x, Eigen::MatrixXd& fjac ) const
-    {
-        const Scalar a = x[0];
-        const Scalar b = x[1];
-        const Scalar c = x[2];
-
-        for ( int i = 0; i < cloud_.points.rows(); ++i )
-        {
-            const InputType p = cloud_.points.row( i ).transpose();
-            const ValueType denominator = p.transpose() * B_ * p;
-            const JacobianType dfda = 2 * ( B_ * p ) * ( p.transpose() * A_ * p ) / ( denominator * denominator );
-            const JacobianType dfdb = -2 * ( B_ * p ) * ( p.transpose() * A_ * p ) / ( denominator * denominator * denominator );
-            const JacobianType dfdc = JacobianType::Zero();
-
-            fjac( i, 0 ) = dfda( 0 );
-            fjac( i, 1 ) = dfdb( 0 );
-            fjac( i, 2 ) = dfdc( 0 );
-        }
-
-        return 0;
-    }
-
-    PointCloud cloud_;
-    InputType A_;
-    InputType B_;
-};
-
-
-
-#endif 
-
-
-
-#if 0
-
-
-struct ConicalFittingFunctor : Functor<double>
-{
-    ConicalFittingFunctor( const Vector3d& vertex, const Vector3d& axis, double openingAngle, const MatrixXd& points )
-        : Functor<double>( points.size() ), vertex_( vertex ), axis_( axis.normalized() ), openingAngle_( openingAngle ), points_( points )
-    {}
-
-    int operator()( const VectorXd& x, VectorXd& fvec ) const
-    {
-        Vector3d apex = vertex_ + x.head( 3 );
-
-        double k = std::tan( openingAngle_ );
-        Vector3d dir( x( 3 ), x( 4 ), x( 5 ) );
-
-        for ( int i = 0; i < points_.rows(); ++i )
-        {
-            const Vector3d& p = points_.row( i );
-            Vector3d v = apex - p;
-            fvec( i ) = v.dot( axis_ ) - v.norm() * k * dir.dot( v );
-        }
-
-        return 0;
-    }
-
-private:
-    Vector3d vertex_;
-    Vector3d axis_;
-    double openingAngle_;
-    MatrixXd points_;
-};
-
-#endif 
-
-
-
-template <typename T>
-//using T = float;
-class Cone3
-{
-public:
-    Cone3()
-    {}
-
-
-    Cone3( const Line3<T>& inAxis, T inAngle, T inLength )
-        :
-        position( inAxis ),
-        angle( inAngle ),
-        length( inLength )
-    {}
-
-    inline MR::Vector3<T>& center( void )
-    {
-        return position.p;
-    }
-    inline MR::Vector3<T>& direction( void )
-    {
-        return position.d;
-    }
-    inline MR::Vector3<T>& apex( void )
-    {
-        return center();
-    }
-
-
-    MR::Line3<T> position; // the combination of the apex of the cone and the direction of its main axis in space
-    T angle = 0;
-    T length = 0;
-};
-
-
-
-
-
-
-//using T = float;
-template <typename T>
-struct ConeFittingFunctor
-{
-    using Scalar = T;
-    using InputType = Eigen::Matrix<T, Eigen::Dynamic, 1>;
-    using InputType2 = Eigen::Vector<T, Eigen::Dynamic>;
-    using ValueType = Eigen::Matrix<T, Eigen::Dynamic, 1>;
-    using JacobianType = Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic>;
-
-    std::vector <Eigen::Vector3<T>> points;
-
-    int inputs() const
-    {
-        return 6;
-    }
-    int values() const
-    {
-        return static_cast< int > ( points.size() );
-    }
-
-    // F[i](V,W) = D^T * (I - W * W^T) * D, D = V - X[i], P = (V,W)
-    int operator()( const InputType& x, ValueType& F ) const
-    {
-        Eigen::Vector3<T> V = x.head<3>();
-        Eigen::Vector3<T> W = x.segment<3>( 3 );
-
-        for ( int i = 0; i < points.size(); ++i )
-        {
-            Eigen::Vector3<T> delta = V - points[i];
-            T deltaDotW = delta.dot( W );
-            F[i] = delta.squaredNorm() - deltaDotW * deltaDotW;
-        }
-        return 0;
-    }
-
-
-
-    int df( const InputType& x, JacobianType& J ) const
-    {
-        Eigen::Vector3<T> V = x.head<3>();
-        Eigen::Vector3<T> W = x.segment<3>( 3 ).normalized();
-
-        for ( int i = 0; i < points.size(); ++i )
-        {
-            const Eigen::Vector3<T>& P = points[i];
-            Eigen::Vector3<T> D = ( V - P );
-            T PW = D.dot( W );
-            Eigen::Vector3<T> PVW = V - P - PW * W;
-            Eigen::Vector3<T> PWD = PW * D;
-
-            // Derivative of f with respect to the components of vertex V
-            J( i, 0 ) = 2 * PVW.x();
-            J( i, 1 ) = 2 * PVW.y();
-            J( i, 2 ) = 2 * PVW.z();
-
-            // Derivative of f with respect to the components of the vector W
-            J( i, 3 ) = -2 * PWD.x();
-            J( i, 4 ) = -2 * PWD.y();
-            J( i, 5 ) = -2 * PWD.z();
-        }
-        return 0;
-    }
-
-};
-
-
-#if 0 
-mFFunction = [this] ( GVector<T> const& P, GVector<T>& F )
-{
-    Vector3<T> V = { P[0], P[1], P[2] };
-    Vector3<T> W = { P[3], P[4], P[5] };
-    for ( int32_t i = 0; i < mNumPoints; ++i )
-    {
-        Vector3<T> delta = V - mPoints[i];
-        T deltaDotW = Dot( delta, W );
-        F[i] = Dot( delta, delta ) - deltaDotW * deltaDotW;
-    }
-};
-#endif
-//using T = float;
-template <typename T>
-class Cone3Approximation
-{
-public:
-    Cone3Approximation()
-    {
-
-
-
-
-    }
-
-
-    // The parameters coneVertex, coneAxis and coneAngle are in/out
-    // variables. The caller must provide initial guesses for these.
-    // The function estimates the cone parameters and returns them. See
-    // GaussNewtonMinimizer.h for a description of the least-squares
-    // algorithm and the parameters that it requires. (The file
-    // LevenbergMarquardtMinimizer.h directs you to the Gauss-Newton
-    // file to read about the parameters.)
-
-            //size_t maxIterations, T updateLengthTolerance, T errorDifferenceTolerance,
-//T lambdaFactor, T lambdaAdjust, size_t maxAdjustments,
-
-    void fitParamsToCone( Eigen::Vector<T, Eigen::Dynamic>& fittedParams, Cone3<T>& cone )
-    {
-        cone.apex().x = fittedParams[0];
-        cone.apex().y = fittedParams[1];
-        cone.apex().z = fittedParams[2];
-
-        cone.direction().x = fittedParams[3];
-        cone.direction().y = fittedParams[4];
-        cone.direction().z = fittedParams[5];
-
-        cone.direction().normalized();
-    }
-    void coneToFitParams( Cone3<T>& cone, Eigen::Vector<T, Eigen::Dynamic>& fittedParams )
-    {
-        // The fittedParams guess for the cone vertex.
-        fittedParams[0] = cone.apex().x;
-        fittedParams[1] = cone.apex().y;
-        fittedParams[2] = cone.apex().z;
-
-
-        // The initial guess for the weighted cone axis.
-        T coneCosAngle = std::cos( cone.angle );
-
-        fittedParams[3] = cone.direction().x / coneCosAngle;
-        fittedParams[4] = cone.direction().y / coneCosAngle;
-        fittedParams[5] = cone.direction().z / coneCosAngle;
-
-        //fittedParams[3] = cone.direction().x;
-        //fittedParams[4] = cone.direction().y;
-        //fittedParams[5] = cone.direction().z;
-
-
-    }
-
-
-
-    void solve( const std::vector<MR::Vector3<T>>& points,
-            Cone3<T>& cone  /*, bool useConeInputAsInitialGuess = false */ )
-
-    {
-        ConeFittingFunctor<T> coneFittingFunctor;
-        Eigen::LevenbergMarquardt<ConeFittingFunctor<T>, T> lm( coneFittingFunctor );
-        MR::Vector3<T>& coneVertex = cone.apex();
-        MR::Vector3<T>& coneAxis = cone.direction();
-        T& coneAngle = cone.angle;
-
-        /*
-        if ( useConeInputAsInitialGuess )
-        {
-            Normalize( coneAxis );
-        }
-        else
-        */
-        {
-            ComputeInitialCone( points, coneVertex, coneAxis, coneAngle );
-            //           ComputeInitialCone( points, cone.apex(), cone.direction(), cone.angle() );
-        }
-
-        Eigen::VectorX<T> fittedParams( 6 );
-        coneToFitParams( cone, fittedParams );
-        [[maybe_unused]] Eigen::LevenbergMarquardtSpace::Status result = lm.minimize( fittedParams );
-
-        // Looks like a bug in Eigen. Eigen::LevenbergMarquardtSpace::Status have error codes only. Not return value for Success minimization. 
-        // So just log status 
-#if 0
-        if ( result == Eigen::LevenbergMarquardtSpace::Status::Success )
-        {
-            std::cout << "Converged to solution: " << fittedParams.transpose() << std::endl;
-        }
-        else
-        {
-            std::cout << "Levenberg-Marquardt optimization failed." << std::endl;
-        }
-#endif 
-
-        //auto result = minimizer( initial, maxIterations, updateLengthTolerance,
-        //    errorDifferenceTolerance, lambdaFactor, lambdaAdjust, maxAdjustments );
-
-        // No test is made for result.converged so that we return some
-        // estimates of the cone. The caller can decide how to respond
-        // when result.converged is false.
-
-        fitParamsToCone( fittedParams, cone );
-
-
-        // We know that coneCosAngle will be nonnegative. The std::min
-        // call guards against rounding errors leading to a number
-        // slightly larger than 1. The clamping ensures std::acos will
-        // not return a NaN.
-        T const one_v = static_cast< T >( 1 );
-        auto coneCosAngle = std::min( one_v / coneAxis.length(), one_v );
-        cone.angle = std::acos( coneCosAngle );
-
-        cone.length = 5.0f; // TODO solve cone length as smth.
-
-        return;
-    }
-
-
-
-private:
-
-
-    // Function for finding the best approximation of a straight line in general form y = a*x + b
-    void findBestFitLine( const std::vector<MR::Vector2<T>>& xyPairs, T& lineA, T& lineB, MR::Vector2<T>* avg = nullptr )
-    {
-        auto numPoints = xyPairs.size();
-        Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic> A( numPoints, 2 );
-        Eigen::Vector<T, Eigen::Dynamic>  b( numPoints );
-
-        for ( auto i = 0; i < numPoints; ++i )
-        {
-            A( i, 0 ) = xyPairs[i].x;  // x-coordinate of the i-th point
-            A( i, 1 ) = 1.0;           // constant 1.0 for dummy term b
-            b( i ) = xyPairs[i].y;     // y-coordinate of the i-th point
-            if ( avg )
-                *avg = *avg + xyPairs[i];
-        }
-        if ( avg )
-        {
-            *avg = *avg / static_cast < T > ( xyPairs.size() );
-        }
-        // Solve the system of equations Ax = b using the least squares method
-        Eigen::Matrix<T, Eigen::Dynamic, 1> x = A.bdcSvd( Eigen::ComputeThinU | Eigen::ComputeThinV ).solve( b );
-
-        lineA = x( 0 );
-        lineB = x( 1 );
-
-        if ( avg )
-        {
-            *avg = *avg / static_cast < T > ( xyPairs.size() );
-            avg->y = lineA * avg->x + lineB;
-        }
-
-    }
-
-    /*
-    void findLineApproximation( const std::vector<Vector2<T>>& xyPairs, T& xAvr, T& yAvr, T& a )
-    {
-        using Matrix = Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic>;
-        using Vector = Eigen::Matrix<T, Eigen::Dynamic, 1>;
-
-        auto  m = xyPairs.size();
-
-        // Шаг 1: Вычисление средних значений
-        xAvr = 0;
-        yAvr = 0;
-        for ( const auto& pair : xyPairs )
-        {
-            xAvr += pair.x;
-            yAvr += pair.y;
-        }
-        xAvr /= m;
-        yAvr /= m;
-
-        // Создание матрицы A и столбца b
-        Matrix A( m, 2 );
-        Vector b( m );
-
-        // Шаги 2-4: Заполнение матрицы A и столбца b
-        for ( int i = 0; i < m; ++i )
-        {
-            A( i, 0 ) = xyPairs[i].x - xAvr;
-            A( i, 1 ) = 1;
-            b( i ) = xyPairs[i].y - yAvr;
-        }
-
-        // Шаг 5: Решение линейной системы
-        Vector x = A.fullPivLu().solve( b );
-
-        // Шаг 6: Получение значений a и c
-        a = x( 0 );
-    }
-    */
-
-
-    void ComputeInitialCone( const std::vector<MR::Vector3<T>>& points, MR::Vector3<T>& coneVertex, MR::Vector3<T>& coneAxis, T& coneAngle )
-    {
-        // Compute the average of the sample points.
-        MR::Vector3<T> center{ 0, 0, 0 };
-        for ( auto i = 0; i < points.size(); ++i )
-        {
-            center += points[i];
-        }
-        center = center / static_cast< T >( points.size() );
-
-        // The cone axis is estimated from ZZTZ (see the PDF).
-        coneAxis = { 0, 0, 0 };
-        for ( auto i = 0; i < points.size(); ++i )
-        {
-            Vector3<T> delta = points[i] - center;
-            coneAxis += delta * MR::dot( delta, delta );
-        }
-        coneAxis = coneAxis.normalized();
-
-        // Compute the signed heights of the points along the cone axis
-        // relative to C. These are the projections of the points onto the
-        // line C+t*U. Also compute the radial distances of the points
-        // from the line C+t*U.
-        std::vector<Vector2<T>> hrPairs( points.size() );
-        T hMin = std::numeric_limits<T>::max(), hMax = -hMin;
-        for ( auto i = 0; i < points.size(); ++i )
-        {
-            MR::Vector3<T> delta = points[i] - center;
-            T h = MR::dot( coneAxis, delta );
-            hMin = std::min( hMin, h );
-            hMax = std::max( hMax, h );
-            Vector3<T> projection = delta - MR::dot( coneAxis, delta ) * coneAxis;
-            T r = projection.length();
-            hrPairs[i] = { h, r };
-        }
-
-        // The radial distance is considered to be a function of height.
-        // Fit the (h,r) pairs with a line:
-        //   r - rAverage = hrSlope * (h - hAverage)
-        MR::Vector2<T> avgPoint;
-        T a, b; // line y=a*x+b
-        findBestFitLine( hrPairs, a, b, &avgPoint );
-        T hAverage = avgPoint.x;
-        T rAverage = avgPoint.y;
-        T hrSlope = a;
-
-        // If U is directed so that r increases as h increases, U is the
-        // correct cone axis estimate. However, if r decreases as h
-        // increases, -U is the correct cone axis estimate.
-        if ( hrSlope < 0 )
-        {
-            coneAxis = -coneAxis;
-            hrSlope = -hrSlope;
-            std::swap( hMin, hMax );
-            hMin = -hMin;
-            hMax = -hMax;
-        }
-
-        // Compute the extreme radial distance values for the points.
-        T rMin = rAverage + hrSlope * ( hMin - hAverage );
-        T rMax = rAverage + hrSlope * ( hMax - hAverage );
-        T hRange = hMax - hMin;
-        T rRange = rMax - rMin;
-
-        // Using trigonometry and right triangles, compute the tangent
-        // function of the cone angle.
-        T tanAngle = rRange / hRange;
-        coneAngle = std::atan2( rRange, hRange );
-
-        // Compute the cone vertex.
-        T offset = rMax / tanAngle - hMax;
-        coneVertex = center - offset * coneAxis;
-    }
-};
-
-
-
-
-
-
-
-
-
-
-
-
 
 }
